+++ snippy/config/source/cli.py
@@ -62,6 +62,7 @@ class Cli(ConfigSourceBase):
         '  --filter REGEXP               filter search result with regexp',
         '  --limit INT                   maximum number of search results',
         '  --sort FIELD                  sort search result based on fields',
+        '  --headers                     print only content headers',
         '  --no-ansi                     remove ANSI characters from output'
     )
     ARGS_MIGRATE = (
@@ -195,6 +196,7 @@ class Cli(ConfigSourceBase):
         search.add_argument('--filter', type=Parser.to_unicode, dest='search_filter', default=argparse.SUPPRESS, help=argparse.SUPPRESS)
         search.add_argument('--limit', type=int, default=Cli.LIMIT_DEFAULT_CLI, help=argparse.SUPPRESS)
         search.add_argument('--sort', nargs='*', type=Parser.to_unicode, default='brief', help=argparse.SUPPRESS)
+        search.add_argument('--headers', action='store_true', default=argparse.SUPPRESS, help=argparse.SUPPRESS)
 
         # migration options
         migrat = parser.add_argument_group(title='migration options', description=Const.NEWLINE.join(Cli.ARGS_MIGRATE))